Added Nightvision and FLIR to Helmet AR
This is a new Helmet AR app, that adds Night vision and FLIR to the Helmet AR.

This is basically the Workshop version from Spy back in 2014. Fixed it up a bit. Used with permission from Spy.

Plan on changing things up a bit in the future. This is just the initial release.